+++ CHANGELOG.md
@@ -10,12 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [7.23.0] - 2025-05-29
 ### Added
 - [JUnit Platform Engine, TestNG] Remove framework elements from `UndefinedStepException` stacktrace ([#3002](https://github.com/cucumber/cucumber-jvm/pull/3002) M.P. Korstanje)
 - [JUnit Platform Engine] Add `surefire` naming strategy ([#3003](https://github.com/cucumber/cucumber-jvm/pull/3003) M.P. Korstanje)
 
 ### Changed
-- [JUnit Platform Engine] Use `number-and-pickle-if-parameterized ` example naming strategy by default ([#3004](https://github.com/cucumber/cucumber-jvm/pull/3004) M.P. Korstanje)
+- [JUnit Platform Engine] Use `number-and-pickle-if-parameterized` example naming strategy by default ([#3004](https://github.com/cucumber/cucumber-jvm/pull/3004) M.P. Korstanje)
 
 ## [7.22.2] - 2025-05-12
 ### Changed
@@ -409,7 +411,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Weld] Removed `cucumber-weld` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 - [Needle] Removed `cucumber-needled` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.22.2...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.23.0...HEAD
+[7.23.0]: https://github.com/cucumber/cucumber-jvm/compare/v7.22.2...v7.23.0
 [7.22.2]: https://github.com/cucumber/cucumber-jvm/compare/v7.22.1...v7.22.2
 [7.22.1]: https://github.com/cucumber/cucumber-jvm/compare/v7.22.0...v7.22.1
 [7.22.0]: https://github.com/cucumber/cucumber-jvm/compare/v7.21.1...v7.22.0